requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
autobahn==22.12.1
Automat==22.10.0
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==38.0.4
daphne==4.0.0
Django==4.1.4
django-cors-headers==4.0.0
django-rest-framework==0.1.0
django-simple-history==3.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.4
phonenumbers==8.13.2
Pillow==9.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.1.0
pytz==2022.6
redis==4.4.0
requests==2.31.0
requests-toolbelt==1.0.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
urllib3==2.0.2
yapf==0.32.0
zope.interface==5.5.2
